# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Pocassist Api",
        "title": "Pocassist Api",
        "contact": {},
        "version": "v0.4.0"
    },
    "paths": {
        "/api/v1/poc/": {
            "get": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "plugin list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Pagesize",
                        "name": "pagesize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "affectsField",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "enableField",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "新增",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "plugin add",
                "parameters": [
                    {
                        "description": "plugin",
                        "name": "plugin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rule.Plugin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/poc/download/": {
            "post": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "下载yaml",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "download yaml",
                "parameters": [
                    {
                        "description": "json_poc",
                        "name": "run",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/plugin.DownloadSwigger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/poc/run/": {
            "post": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "运行",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "plugin run",
                "parameters": [
                    {
                        "description": "run",
                        "name": "run",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/plugin.RunSwigger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/poc/upload/": {
            "post": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "上传yaml",
                "consumes": [
                    "multipart/form-data",
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "upload yaml",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file",
                        "name": "yaml",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/poc/{id}/": {
            "get": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "plugin detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "更新",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "plugin update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "plugin",
                        "name": "plugin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rule.Plugin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "删除",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "plugin delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/result/": {
            "get": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Result"
                ],
                "summary": "result list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Pagesize",
                        "name": "pagesize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "taskField",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "vulField",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/result/{id}/": {
            "delete": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "删除",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Result"
                ],
                "summary": "result delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "msg.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "plugin.DownloadSwigger": {
            "type": "object",
            "properties": {
                "json_poc": {
                    "$ref": "#/definitions/rule.Plugin"
                }
            }
        },
        "plugin.RunSwigger": {
            "type": "object",
            "properties": {
                "affects": {
                    "type": "string"
                },
                "json_poc": {
                    "$ref": "#/definitions/rule.Plugin"
                },
                "target": {
                    "type": "string"
                },
                "vul_id": {
                    "type": "string"
                }
            }
        },
        "rule.Detail": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "rule.Plugin": {
            "type": "object",
            "properties": {
                "affects": {
                    "description": "影响类型  dir/server/param/url/content",
                    "type": "string"
                },
                "enable": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "jsonPoc": {
                    "description": "json规则",
                    "$ref": "#/definitions/rule.Poc"
                },
                "vulId": {
                    "description": "漏洞编号",
                    "type": "string"
                }
            }
        },
        "rule.Poc": {
            "type": "object",
            "properties": {
                "detail": {
                    "$ref": "#/definitions/rule.Detail"
                },
                "groups": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/rule.Rule"
                        }
                    }
                },
                "name": {
                    "type": "string"
                },
                "params": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rule.Rule"
                    }
                },
                "set": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "rule.Rule": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "expression": {
                    "type": "string"
                },
                "follow_redirects": {
                    "type": "boolean"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "search": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}